# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Personal Finance Tracker API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /transactions:
    get:
      operationId: getTransactions
      tags: [transactions]
      summary: List transactions
      parameters:
        - name: month
          in: query
          schema:
            type: ["string", "null"]
          description: Filter by month (YYYY-MM)
        - name: categoryId
          in: query
          schema:
            type: ["integer", "null"]
        - name: type
          in: query
          schema:
            type: ["string", "null"]
            enum: [income, expense, transfer, null]
        - name: search
          in: query
          schema:
            type: ["string", "null"]
        - name: limit
          in: query
          schema:
            type: ["integer", "null"]
        - name: offset
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: List of transactions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionListResponse"
    post:
      operationId: createTransaction
      tags: [transactions]
      summary: Create a transaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTransactionInput"
      responses:
        "201":
          description: Created transaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /transactions/{id}:
    get:
      operationId: getTransaction
      tags: [transactions]
      summary: Get a transaction
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Transaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateTransaction
      tags: [transactions]
      summary: Update a transaction
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTransactionInput"
      responses:
        "200":
          description: Updated transaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteTransaction
      tags: [transactions]
      summary: Delete a transaction
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    Transaction:
      type: object
      properties:
        id:
          type: integer
        date:
          type: string
          format: date
        description:
          type: string
        amount:
          type: number
        type:
          type: string
          enum: [income, expense, transfer]
        categoryId:
          type: ["integer", "null"]
        categoryName:
          type: ["string", "null"]
        categoryColor:
          type: ["string", "null"]
        categoryIcon:
          type: ["string", "null"]
        tags:
          type: array
          items:
            type: string
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - date
        - description
        - amount
        - type
        - tags
        - createdAt
    TransactionListResponse:
      type: object
      properties:
        transactions:
          type: array
          items:
            $ref: "#/components/schemas/Transaction"
        total:
          type: integer
        totalIncome:
          type: number
        totalExpenses:
          type: number
      required:
        - transactions
        - total
        - totalIncome
        - totalExpenses
    CreateTransactionInput:
      type: object
      properties:
        date:
          type: string
          format: date
        description:
          type: string
        amount:
          type: number
        type:
          type: string
          enum: [income, expense, transfer]
        categoryId:
          type: ["integer", "null"]
        tags:
          type: array
          items:
            type: string
        notes:
          type: ["string", "null"]
      required:
        - date
        - description
        - amount
        - type
    UpdateTransactionInput:
      type: object
      properties:
        date:
          type: string
          format: date
        description:
          type: string
        amount:
          type: number
        type:
          type: string
          enum: [income, expense, transfer]
        categoryId:
          type: ["integer", "null"]
        tags:
          type: array
          items:
            type: string
        notes:
          type: ["string", "null"]
...